+++ src/cli.rs
@@ -38,7 +38,7 @@ It creates an empty changelog skeleton if no existing changelog is found as well
 
 #[derive(Args, Debug)]
 pub struct AddArgs {
-    pub number: Option<u16>,
+    pub number: Option<u64>,
     #[arg(short, long)]
     pub yes: bool,
 }
